Removed parallel output
Each character put was really slow when trying to output it to the
parallel port on real hardware (presumably because nothing is listening,
or perhaps due to the ugly delay()ing).  The only thing that used it was
Bochs, so just go ahead and redirect your serial out to a terminal like
a real man/woman.