SMP stacks and initial INT handling
Per-core stacks, pages of which include per-core GDT and Taskstates.
Currently using the same IDT for all cores, and the same
interrupt_handler table.  Also removed the old "sizeof(struct
Trapframe)" from the bootstack_top and smp variety.  Not sure why it was
there, and it works fine without it.