Added shared info/data pages and removed UENVS
Changed the memory mappings and create per-process shared info (RO) and
shared data (RW) pages at specific locations in the address space.  Plus
assorted changes to pmap and env to handle the loss of UENVS and the new
mappings.  Also had to adjust userspace slightly to use the new labels.
Will need to sort the TODOs when we create actual structures to go in
each shared page.