smp_call backend removed, other async tweaks
Merged the back and frontend to just one checklist, since they both
needed to be waited on.  Added some other checks to prevent trying to
grab a handler when none are available and will never become available
(like if they are all being waited on and the caller could be the
waiter).  Uses a global counter to track the number of outstanding
calls.  Also has support for commit_checklist to preemptively abort if
it knows it will take a while (locked or not clear).