Async call responses
Pass back some sort of response from the syscalls within an async call.
Could consider passing in a function that does the aggregation of
responses, since it may be syscall specific.

Also, process_syscalls no longer resets the machine when trying to
process the syscalls of a process that's been freed.  Though there is no
sync protection with that yet.