Async error handling
Receives error return values and passes them back up the userspace
stack.  It's up to the top levels to deal with them.  cprintf currently
does not, and null does.

Also adjusted measure_ to handle an error (presumably RING_FULL), and
not try that iteration again.  The numbers may be a little odd from it
if the error doesn't happen on the first outer iteration.

Also, for any real measurements, note the final parameter of
process_generic_syscalls.