Cache-buster syscall, sync and async varieties
Just grabs a lock and writes a value to a certain number of entries to
an array in kernel space.  Right now, the array is sitting at
0xd0000000, which is far from being allocated for anything.

Also added a wrapper function to get both an async and a sys desc, which
seems to be the common case for a lot of syscalls.