Added some function to read the tsc with start/stop semantics
In addition to these functions, there is a training function that figures out what the correct
overhead of making the calls is so that it can be factored out before returning the time
difference.  This has not been tested on real hardware yet, and is pretty flaky in my
bochs-inside-vmware environment.  Logic seems sound though...