Merged the timing and measurement stuff together and cleaned it up a bit
This involved moving some things around as well as renaiming some things.  I'm still not
happy with the name timer.h/c because these names imply fucntionality such as being able to
set an interrrupt based timer, etc.  But we can change this later if we want to.  I also made
some small modifications to finally get rid of all remnants of the "jos" tainted throughout
our code.  Additionally, there was some reorganization of who includes what header files so as
to be consistent with what is really needed by any individual file.