Beginnings of getting user apps to compile against newlib
I've compiled newlib outside of this source tree with the -nostdinc option for linking so that none
of the libraries from my local machine got bundled with it as it was compiled.  It is installed in
the newlib directory along with some files for providing the necessary libc wrappers for our
platform.  The bodies of the functions found in libc_wrappers.c should be filled in to define how
these functions interact with ROS.  Right now, noe of them do the right thing :)  We will fill them
in as we go along...