SMP boot cleanup
Deallocates the *page table* used by the smp_boot mapping, which was
being copied into every processes address space, which caused problems.

Need to take a look at UTOP and ULIM and add a check in env_vm_setup to
check for these things in the future.  Do this when gutting UPAGES.