Fixed print macros in stdio.h, fixed bug in readline.c, expanded matrix.c
Fixed a bug in the arguments of the printk and printd macros in stdio.h. This bug
caused strange behavior in user land apps, such as the arguments to sbrk being
trashed.

Corrected the behavior of readline() to detect errors on the return value, not
the character returned, and removed a floating point printf format character.

Expanded matrix.c to include the full help message and execute the correct apps.