Built a basic framework for sending ethernet packets from userland. Switched from measurements to matrix.
Built a syscall framework called sys_eth_write/read that sends and receives packets. It does this by (stupidly) wrapping the data to send with
hard coded ETH/IP/UDP headers, and pushing it out on the NIC. There is currently no mechinism for receiving packets, as userland interrupts are not
working, and those are required for incoming notification. This current version will send properly formatted UDP packets out on the wire.

I also modified newlib_backend to use this new medium, and (for debugging) reenabled the hacks on isatty and fstat to make printf not go over the wire.

I also merged back in the code to switch  to the user enviroment, and run matrix.c.