Fixed bug with sysenter not restoring interrupts. Cleaned up mistakenly left in debug lines in nic driver
Sysenter, somewhat covertly, disables the IF bit in the eflags register upon its invokation. This is
described only in the RTL, not the functional description. The result of this is upon entering userland,
if syscalls were serviced by the sysenter facility, interrupts would be disabled and not come back online.
The fix is to force the IF bit back  on as soon as we get inside the syscall wrapper.

Also removed some debugging stuff from rl8168.c.