Got the File IO tests up and running over ethernet.
This is incredibly hacky. I am simulating UDP/IP at the driver/syscall level because we lack a network stack.
The code in handle_rx_packet() is really really ugly and a total hack. This needs to go away.

I also fixed the tcp/ip/udp checksum flags in the descriptor register. Their offsets are now correct
(and undocummented in the spec sheet). They need to be handled as part of the vlan field instead of command.
Note: The nic is stupid. If you send a UDP packet with the TCP checksum enabled, it sticks 16 bits where
it thinks it should go, right over your data.