Fixed a bug in apic.c for irq's > 7. Rewrote/organized nic driver to use kalloc
Fixed a bug in apic.c that masked out the wrong irq in PIC1 for IRQ's over 7
in PIC2.

Rewrote parts of the realtek driver to use kalloc instead of page_alloc, also
reorganized the driver and header in a better manor. Isolated out the hacky
stuff and cleaned up as much as possible.

This commit will not compile, as I'm about to rebase based on the new master
branch that will have kalloc.