Moved IOAPIC functionality into ioapic.c, added support for ISA IRQs (pit keyboard, etc)
Net result is we can now reroute any IRQ on the PCI or ISA bus, this includes
anything that is listed on the ISA bus (in mptables), such as the PIT,
Keyboard, etc.

Moved ioapic_reroute_irq() into ioapic.c. Also added an ioapic_init() function
that runs through the PIC and ISA structures (discussed below) and creates an
IRQ->IOAPIC mapping, so there is a single source of truth as far as the route
functions are concerned.

Added an ioapic_unroute_irq() that basically masks an irq.

Added ISA IRQ related parsing to MPTables that allows for things like the PIT
to be routed via the IOAPIC. This creates a structure like the one found for
pci that maps IRQ's to ioapic entries. This is used in ioapic_init()

Added ioapic_init() to the init process after pci_init.