Added support for bochs/kvm broken mptables, added pit reroute test, fixed imcr bug
Added support for mismatched pci bus and mptables settings. Specifically, fixed
an issue where bochs and kvm say there is no pci bus in the mptable, but have
a pci bus that is readable. The missing devices seem to show up as isa devices
in the mptable. This code trys to gracefully handle this. It is untested, in
terms of interrupts coming from pci devices in bochs/kvm, as we don't have a
way to generate such interrupts right now (quickly).

Added test_ioapic_pic_reroute() in testing.c that routes the pit to a core,
using the given ioapic_reroute_irq() interface, enables the pit, then unroutes
it using ioapic_unroute_irq().

Fixed a bug where the interupt mode bit mask in the mptable was a 16 bit field
instead of an 8 bit field.