Modified some stuff to cleanup the syscall_server build process
This also involved adding a sandbox directory under the serial_server directory.  All files accessed by ROS will be relative to this sandbox.  There is currently no suport for preventing one from using ".." to break out of the sandbox.  Presumably a real server would run with itself as a user with no priveleges other than to access files in the sandbox directory and nowhere else.