Fixed issues building without ivy and without __NETWORK__
There was an issue where the libc_patch.i file was being included by a pair
of userspace Makefrags, regardless of ivy being set as the compiler. This caused
the build process to fail if ivy was not installed on the system trying to build
the OS. This is now fixed.

I moved a pair of cases in the main syscall switch inside of a __NETWORK__ ifdef.
Previously they were inside of an __i386__ ifdef. This meant an error would occur
if the combination of __i386__ was set without __ NETWORK__.