Active messages using dynamically allocated memory
This should avoid the active message deadlock issues.  This uses the
slab allocator to create new messages, so a core doesn't run out of
slots for amsgs.  You could still run out of memory...

Note that the amsg is allocated on one core and freed on the other.
This really wants a magazine layer or something similar per-core, so
these allocations and frees don't contend on a global cache lock.

We can gut some things in a future commit, such as the sync variety of
send_amsg, and any of the TODO (AMDL)s.

Also note there is something seriously wrong, perhaps just with KVM, in
which an interrupt coming in around the time of proc_startcore can make
someone "forget" they are holding the lock.  Or rather other people spin
waiting for a lock that never gets released (though the guilty party
progresses as if it was normal).  (TODO).