Per-cpu data init
Adds a function to properly initialize the percpu data for every core.
(Core 0 was neglected).  This needs to be called by each core at the end
of smp_boot.  Lets keep the arches separate at this point, in case we
have some arch specific percpu info.

Also does a little test where spawn creates a couple processes, one of
which goes all multi and stuff, and moves to other cores.