Added (temporary) mechanism to pass argc/argv
Obviously, the kernel should not be responsible for passing
args to user programs; that's a job for the runtime system.
However, to run real programs from within the manager, this
is a necessary evil.

I pass argc/argv inside of the procinfo struct.  They
are set using a call to proc_init_argc_argv(), and they
are read out by user programs within parlibmain().  The
latter no longer expects to be passed argc/argv; instead,
it grabs them from procinfo.