Marked lock-required proc functions with __
In general, I'm using __ to denote a somewhat special function, usually
meant for more internal usage.  In the case of these proc functions, you
need to hold the lock before you call it.  In some other cases, the
functions are internal and not meant to be called directly in most cases
(like __spin_lock, or the active message handlers).