Fixed issues with the RealTek network driver and the slab allocator
Previously the RealTek network driver would use kmalloc for allocating
the RX and TX descriptor rings. A requirement of the rings was that
they be 256 byte alligned. We had previously discussed this, and
determined that by default kmalloc would, if asked for X bytes, return
X byte aligned memory. This is no longer the case with the new
slab allocator, so the driver no longer functioned correctly

We now directly request pages, to ensure proper alignment.