MM touchups
Fixes mmap() for x86 - at least for what it's used for now.

Keeps the sys_ functions in syscall.c.  The general model I'd like is
for the sys_ to be a wrapper with basic checks and a callout to a more
complicated function that might get used elsewhere in system.  We don't
do it well yet, but that's the idea.

Also keeps the bits in mman.h in sync with Linux, which might make
things easier.  If we're going to change them back and forth, there
ought to be a reason at least.

Please keep the syscalls in order in syscall.h and syscall.c, and try to
group related calls together.