Made BusyBox work
Minor tweaks to mmap, mostly.

We need a systematic way for the kernel to execute system
calls "anonymously"--for example, to launch the login shell
without having it stored in KFS, we need to be able to
open, read, and close a file.  I made special internal
versions of these, but it's a bit ugly...