x86: Fixes TLS bug causing kernel page faults
When loading the gs register (or any segmentation register), the
hardware will fill in the internal segmentation registers by looking up
the segment, which can page fault.  The kernel was doing this for
unabandoned, yet unmapped, LDTs from old processes when it tried to
return from an interrupt.

Note that the current TLS implementation is vulnerable to a malicious
process mucking with the LDT.