New vcoremap in procinfo
The vcoremap contains new items (not just pcoreids), and it is exposed
via procinfo (along with num_vcores).  -1 no longer means invalid
either; it's more clear to have a bool for that.  Eventually we'll
signal userspace that the vcoremap (or a pcoremap) is being changed with
the seq_ctr.

If it turns out we are exposing too much info, we can move the vcoremap
back and push updates, but I'd rather not do that.