seq counters to protect vcoremap changes
The kernel uses the seq counter when updating vcore-pcore mappings and
the number of vcores.  Userspace may need to check the counter when
trying to perform certain operations.

Also, this introduces a reference seq lock, that is an example of how to
use it.  It relies on a spinlock, which isn't that flexible.