Fixed issue with make not working right on Fedora, if ivycc wasn't installed
On some systems (Fedora confirmed) 'which non_existant_file' will generate
errors instead of simply not returning anything. The result of this is
attempting to do any of the make operations without ivycc installed would
generate a series of errors. This occured even if the system was configured
to use standard gcc.