x86: waits after sending IPIs
Instead of checking before, wait afterwards.  It should provide the same
protection from losing IPIs (since all IPI-making functions adhere to
this), but now the send functions return once the IPI should have made
it to the distant end.  This way, we don't need to wait twice or have a
separate call for an IPI-barrier (like the wmb()s).