Added BCQs to the preempt structs in procdata
This involved duplicating the code for atomic operations (though for
now, only CAS was needed, and that might go away).  On sparc, this means
copying all of the spinlock stuff too.

Also, bcq_struct.h is included with the structure and other things, with
the actual work done in bcq.h.  This is because libc includes
bcq_struct, but doesn't actually do the ops, and we want to limit the
amount of crap in the ros interface (like comp_and_swap).