Added demand paging support
mmap'd pages are now faulted in on demand.  Metadata about paged-out
pages are stored in the unused bits of the PTE when the PTE is invalid.
Metadata comes in the form of a pointer to type pfault_info_t.

The main thing to keep in mind is it's often no longer sufficient to
check user PTEs against PTE_P to determine their state.  Use the
macros PAGE_PRESENT, PAGE_PAGED_OUT, and PAGE_UNMAPPED instead;
exactly one of the three must be true at any time.

On any call to mmap with fd != -1, a new struct file is instantiated.
It is reference-counted, once per mmap'd page.  When a page is faulted
in, or when a paged-out page is freed, the file is decref'd.  Allocating
and freeing a pfault_info_t will automatically incref/decref the associated
file, if any.