Moved brk/sbrk into user-space
glibc now does it with mmap.  BRK_END is used
to demarcate between the maximum break size
and where non-MAP_FIXED mmap starts allocating,
so the heap and mmap'd memory won't collide.

sys_brk still exists for compatibility with newlib;
someone might want to delete this in the future.

Recompile the cross-compiler, as procinfo changed.