Vcore0 can restart in _M mode
When notified or for other reasons that may cause a restart, vcore0 will
come up on a fresh transition stack (not USTACKTOP) with a fresh TLS
region.  In essence, the context from _S mode is just a user thread that
will need to be managed by a scheduler that needs to be invoked from
hart_entry().

Rebuild your cross compiler.