x86: Userspace code to restart ROS TFs
A user scheduler can pop a TF provided by the kernel, and enable
notifications.  This is the way to leave the transition stack / start up
a user context.

TLSs aren't dealt with in particular.  The code will maintain whatever
gs is loaded, which is set per vcore.  The user scheduler will need to
set up the appropriate LDT entry (on x86) for the TLS/TCB of the user
context it's attempting to load.

This doesn't work on sparc yet.  It will just silently fail (and panic
in mhello).