Fixes a null pointer bug in cprintf
Not that it matters much - it would only happen if you passed a 0 to
cprintf/printk, and you should rarely pass a string directly (format
string issues).  If this bug hit, you'd trap and then deadlock on the
output lock.  Looking at this was on the todo list for a long time, but
it was suddenly triggerable with kfunc cprintf.