Makes userspace debuginfo configurable
Glibc will continue to be built with -g, but if the CONFIG_USERDEBUG
flag isn't set, the make process won't compile user/ and tests/ with -g
and will strip the debuginfo from the binaries, which decreases the size
of the programs (and the kernel when using KFS).

Also, this fixes a bug where the USER_CFLAGS were getting clobbered.