First run at an E1000 driver, with various LWIP fixes.
This is the first pass at a functional E1000 driver for both qemu
and the 8 core nehalem box (s142). This specifically will not work
on other hardware, due to a hard coding of the MMIO addr. This is
a temporary measure until we have a way to handle MMIO better.

This also fixes my COMPLETELY BROKEN recursive lock implimentation.
Woops, my bad. I don't know what the hell I was thinking.

This also adds remote reboot functionality to the webserver. Let's
not talk about it.

Conflicts:

	kern/arch/i686/init.c
	user/apps/parlib/httpserver.c
	user/lwip/core/sys_arch.c
	user/lwip/include/arch/sys_arch.h
	user/lwip/include/lwip/lwipopts.h